backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.18
rembg>=2.0.68
numpy>=1.26.0
numpy-stl>=3.1.0
Pillow>=10.4.0
lxml>=5.2.0
mapbox-earcut>=1.0.6
onnxruntime>=1.19.0
scikit-image>=0.24.0
scikit-learn>=1.5.0
